Need further explanation
We have to distinguish between different platform modes and different spaces on the NVDIMM. 
The platform BIOS (server) supports to platform modes. The memory mode also known as 2LM and the App-Direct mode known as 1LM.
With the 2LM mode setting for the platform DRAM will be used as cache . 
With the 1LM mode setting for the platform DRAM will be still direct available for applications.
To switch between modes a reboot is required. 
That’s the only difference between the two platform modes!

The NVDIMMs have to support spaces for the different modes. Similar to partitions. 
E.g. you can reserve ½ of the NVDIMM space for memory mode and the reaming for App-Direct access (App-Direct space). 
If you run the platform in 2LM mode ½ of the NVDIMM size will be available as main memory. All DRAM will be used as cache for the NVRAM.
You can also use the APP-Direct space in memory mode. E.g. mount with DAX option. 

If you run the system in 1LM mode the memory space on the NVDIMM for memory mode is not accessible.
The App-Direct Space is still accessible and you can access the DRAM like normal DRAM.